requirements.txt
python-telegram-bot==20.3
websocket-client==1.6.1
pandas==2.0.3
python-dotenv==1.0.0